+++ cli/internal/orchestrator/semgrep_rules.go
@@ -1,0 +1,6 @@
+package orchestrator
+
+import _ "embed"
+
+//go:embed rules/semgrep-core.yml
+var semgrepCoreRules []byte